Move search UI code to seperate file
